Admin_Backend/requirements.txt
fastapi
uvicorn
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
pydantic